修复导入信息bug

+++ src/main/java/just/learn/service/UserService.java
@@ -1,7 +1,7 @@
 package just.learn.service;
 
-import just.learn.entity.User;
 import just.learn.entity.PageQueryBean;
+import just.learn.entity.User;
 import just.learn.vo.QueryCondition;
 
 public interface UserService {
@@ -9,11 +9,11 @@ public interface UserService {
 
     public User insert(User user);
 
-    public boolean delete(Integer id);
+    public boolean delete(Long id);
 
     public int update(User user);
 
-    public User getById(Integer id);
+    public User getById(Long id);
 
     PageQueryBean getLimitObjects(QueryCondition condition);
 //public   List<User> getAll();
